print sysinfo at the end of test run
